+++ src/adapters.ts
@@ -84,7 +84,10 @@ class VSCodeEditor implements Editor {
     return {
       theme: configuration.get("theme"),
       fontFamily: configuration.get("fontFamily"),
-      useMdPreview: configuration.get<boolean>("useMdPreview", false)
+      previewMarkdownFiles: configuration.get<boolean>(
+        "previewMarkdownFiles",
+        false
+      )
     };
   }