+++ src/domain.test.ts
@@ -33,7 +33,7 @@ describe("start", () => {
     const configuration = {
       theme: "A custom theme",
       fontFamily: "Helvetica",
-      useMdPreview: false
+      previewMarkdownFiles: false
     };
     jest.spyOn(editor, "getConfiguration").mockReturnValue(configuration);
 
@@ -45,7 +45,7 @@ describe("start", () => {
     expect(settings["terminal.integrated.fontFamily"]).toBe(
       configuration.fontFamily
     );
-    expect(settings["workbench.useMdPreview"]).toBe(false);
+    expect(settings["workbench.previewMarkdownFiles"]).toBe(false);
   });
 
   it("should close all editor tabs", async () => {
@@ -205,7 +205,7 @@ class FakeEditor implements Editor {
   showMessage() {}
 
   getConfiguration(): Configuration {
-    return { theme: null, fontFamily: null, useMdPreview: false };
+    return { theme: null, fontFamily: null, previewMarkdownFiles: false };
   }
 }