+++ src/settings.ts
@@ -46,5 +46,5 @@ export const settings = {
   "workbench.editor.enablePreview": false,
 
   "//l": "Use the Markdown Preview pane to view Markdown files",
-  "workbench.useMdPreview": false
+  "slides.previewMarkdownFiles": true
 };